remove fluff
