+++ packages/types/src/extrinsic/v5/GeneralExt.spec.ts
@@ -7,7 +7,7 @@ import rpcMetadata from '@polkadot/types-support/metadata/static-substrate';
 
 import { TypeRegistry } from '../../create/index.js';
 import { Metadata } from '../../metadata/index.js';
-import { GeneralExt } from './index.js';
+import { GeneralExtrinsic } from './GeneralExtrinsic.js';
 
 const registry = new TypeRegistry();
 const metadata = new Metadata(registry, rpcMetadata);
@@ -18,7 +18,7 @@ describe('GeneralExt', (): void => {
   const extrinsic = '0xe90145006500000000000c01000002000000fc39f7510a4c591e688532a6df54e856a77e92aee8a2372f0f194eea11c487393d5c33fd9c2370dfcb4941c7fe85fc19c31e5c9b0bd5ecb74d3923947c9c5ccf00060000d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d0700e40b5402';
 
   it('Can decode a general extrinsic', (): void => {
-    const genExt = new GeneralExt(registry, extrinsic);
+    const genExt = new GeneralExtrinsic(registry, extrinsic);
 
     expect(genExt.version).toEqual(69);
     expect(genExt.transactionExtensionVersion.toNumber()).toEqual(0);
@@ -41,7 +41,7 @@ describe('GeneralExt', (): void => {
       tip: '0x00000000000000000000000000000000',
       transactionVersion: '0x00000002'
     };
-    const genExt = new GeneralExt(registry, {
+    const genExt = new GeneralExtrinsic(registry, {
       payload
     });
